WIP: Add index.Clear tests (#19)
