--- internal/recordrepository/clear_test.go
@@ -1,1 +1,0 @@
-package recordrepository_test
