+++ internal/server/indexhandler/clear.go
@@ -1,0 +1,31 @@
+package indexhandler
+
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	"github.com/ashep/go-apperrors"
+	"github.com/bufbuild/connect-go"
+
+	proto "github.com/ashep/ujds/sdk/proto/ujds/index/v1"
+)
+
+func (h *Handler) Clear(
+	ctx context.Context,
+	req *connect.Request[proto.ClearRequest],
+) (*connect.Response[proto.ClearResponse], error) {
+	err := h.repo.Clear(ctx, req.Msg.Index)
+
+	switch {
+	case errors.As(err, &apperrors.InvalidArgError{}):
+		return nil, connect.NewError(connect.CodeInvalidArgument, err)
+	case err != nil:
+		c := h.now().Unix()
+		h.l.Error().Err(err).Str("proc", req.Spec().Procedure).Int64("err_code", c).Msg("index repo clear failed")
+
+		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("err_code: %d", c))
+	}
+
+	return connect.NewResponse(&proto.ClearResponse{}), nil
+}
