+++ internal/server/recordhandler/handler.go
@@ -19,7 +19,6 @@ type recordRepo interface {
 	Push(ctx context.Context, indexID uint, schema []byte, records []model.Record) error
 	Get(ctx context.Context, index string, id string) (model.Record, error)
 	GetAll(ctx context.Context, index string, since time.Time, cursor uint64, limit uint32) ([]model.Record, uint64, error)
-	Clear(ctx context.Context, index string) error
 }
 
 type Handler struct {
